// src/types/element.d.ts
interface IPPTBaseElement {
    id: string;
    name: string;
    left: number;
    top: number;
    groupId?: string;
    width: number;
    height: number;
}

/**
 * 元素阴影
 *
 * h: 水平偏移量
 *
 * v: 垂直偏移量
 *
 * blur: 模糊程度
 *
 * color: 阴影颜色
 */
export interface IPPTElementShadow {
    h: number;
    v: number;
    blur: number;
    color: string;
}

/**
 * 元素边框
 *
 * style?: 边框样式（实线或虚线）
 *
 * width?: 边框宽度
 *
 * color?: 边框颜色
 */
export interface IPPTElementOutline {
    style?: "dashed" | "solid";
    width?: number;
    color?: string;
    opacity?: number;
}

/**
 * 元素填充色
 *
 * opacity?: 透明度
 *
 * color?: 填充色颜色
 */
export interface IPPTElementFill {
    color?: string;
    opacity?: number;
}
// ...
export interface ImageElementFilters {
    blur?: string;
    brightness?: string;
    contrast?: string;
    grayscale?: string;
    saturate?: string;
    "hue-rotate"?: string;
    opacity?: string;
}

/**
 * 图片裁剪
 *
 * range: 裁剪范围，例如：[[10, 10], [90, 90]] 表示裁取原图从左上角 10%, 10% 到 90%, 90% 的范围
 *
 * shape: 裁剪形状
 */
export interface ImageElementClip {
    range: [[number, number], [number, number]];
    shape: string;
}

/**
 * 图片元素
 *
 * type: 元素类型 (image)
 *
 * fixedRatio: 固定图片宽高比例
 *
 * rotate: 旋转角度
 *
 * streach: 图片显示形式 0 缩放 1 拉伸
 *
 * src: 图片内容
 *
 * filters?: 图片滤镜
 *
 * clip?: 裁剪信息
 *
 * flipH?: 水平翻转
 *
 * flipV?: 垂直翻转
 *
 * opacity?: 不透明度，默认100（%）
 *
 * outline?: 边框
 *
 * shadow?: 阴影
 *
 * fill?: 填充色
 */
export interface IPPTImageElement extends IPPTBaseElement {
    type: "image";
    fixedRatio: boolean;
    rotate: number;
    streach: 0 | 1;
    src: string;
    filters?: ImageElementFilters;
    clip?: ImageElementClip;
    flipH?: 1 | -1;
    flipV?: 1 | -1;
    opacity?: number;
    outline?: IPPTElementOutline;
    shadow?: IPPTElementShadow;
    fill?: IPPTElementFill;
}
// ...
export interface IPPTLatexElement extends IPPTBaseElement {
    type: "latex";
    src: string;
    text: string;
    rotate: number;
    streach: number;
    flipH?: 1 | -1;
    flipV?: 1 | -1;
    opacity?: number;
    outline?: IPPTElementOutline;
    shadow?: IPPTElementShadow;
    fill?: IPPTElementFill;
}

export type ChartType = "bar" | "line" | "pie" | "funnel";

export interface ChartData {
    labels: string[];
    legends: string[];
    series: number[][];
}

/**
 * 图表元素
 *
 * type: 元素类型（chart）
 *
 * fill?: 填充色
 *
 * chartType: 图表基础类型 bar line pie funnel
 *
 * data: 图表数据
 *
 * axisTransformation: 坐标轴转换
 *
 * src: 图表图片
 *
 * legend?: 图例/位置
 *
 * title?: 标题
 *
 * stack?: 柱状图堆叠显示
 *
 * flipH?: 水平翻转
 *
 * flipV?: 垂直翻转
 *
 * opacity?: 不透明度，默认100（%）
 *
 * outline?: 边框
 *
 * shadow?: 阴影
 *
 * fill?: 填充色
 */
export interface IPPTChartElement extends IPPTBaseElement {
    fixedRatio: boolean;
    type: "chart";
    chartType: ChartType;
    data: ChartData;
    axisTransformation: boolean;
    src: string;
    legend?: "" | "top" | "bottom";
    title?: string;
    stack?: boolean;
    rotate: number;
    streach: number;
    flipH?: 1 | -1;
    flipV?: 1 | -1;
    opacity?: number;
    outline?: IPPTElementOutline;
    shadow?: IPPTElementShadow;
    fill?: IPPTElementFill;
}